+++ projects/gateway2/admin/admin_suite_test.go
@@ -1,4 +1,4 @@
-package iosnapshot
+package admin
 
 import (
 	"testing"
@@ -7,7 +7,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func TestIoSnapshot(t *testing.T) {
+func TestAdmin(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "IOSnapshot Suite")
+	RunSpecs(t, "Admin Suite")
 }
