+++ projects/gateway2/admin/redact.go
@@ -1,7 +1,6 @@
-package iosnapshot
+package admin
 
 import (
-	gloov1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -32,24 +31,6 @@ func redactKubeSecretData(obj client.Object) {
 	// Also need to check for kubectl apply, last applied config.
 	// Secret data can be found there as well if that's how the secret is created
 	redactAnnotations(secret.GetAnnotations())
-}
-
-// redactGlooSecretData modifies the secret to remove any sensitive information
-// The structure of a Secret in Gloo Gateway does not lend itself to easily redact data in different places.
-// As a result, we perform a primitive redaction method, where we maintain the metadata, and remove the entire spec
-func redactGlooSecretData(element *gloov1.Secret) {
-	element.Kind = nil
-
-	redactAnnotations(element.GetMetadata().GetAnnotations())
-}
-
-// redactGlooArtifactData modifies the artifact to remove any sensitive information
-func redactGlooArtifactData(element *gloov1.Artifact) {
-	for k := range element.GetData() {
-		element.GetData()[k] = redactedString
-	}
-
-	redactAnnotations(element.GetMetadata().GetAnnotations())
 }
 
 // redactGlooResourceMetadata modifies the metadata to remove any sensitive information
