+++ projects/gateway2/setup/utils/startfuncs.go
@@ -1,0 +1,8 @@
+package utils
+
+import (
+	"context"
+)
+
+// StartFunc is a function that will be called when the kgateway process runs
+type StartFunc func(ctx context.Context) error
