+++ test/helpers/kube_dump.go
@@ -20,7 +20,7 @@ import (
 	"github.com/solo-io/gloo/pkg/utils/kubeutils/kubectl"
 	"github.com/solo-io/gloo/pkg/utils/kubeutils/portforward"
 	"github.com/solo-io/gloo/pkg/utils/requestutils/curl"
-	"github.com/solo-io/gloo/projects/gloo/pkg/servers/admin"
+	"github.com/solo-io/gloo/projects/gateway2/admin"
 	"github.com/solo-io/go-utils/threadsafe"
 )
